Fix some import cycles
