+++ convex/aiTown/inputs.ts
@@ -2,10 +2,11 @@ import { ObjectType } from 'convex/values';
 import { playerInputs } from './player';
 import { conversationInputs } from './conversation';
 import { agentInputs } from './agentInputs';
-import { MutationCtx } from '../_generated/server';
-import { Id } from '../_generated/dataModel';
-import { engineInsertInput } from '../engine/abstractGame';
 
+// It's easy to hit circular dependencies with these imports
+if (playerInputs === undefined || conversationInputs === undefined || agentInputs === undefined) {
+  throw new Error("Input map is undefined, check if there's a circular import.");
+}
 export const inputs = {
   ...playerInputs,
   // Inputs for the messaging layer.
@@ -21,19 +22,3 @@ export type InputReturnValue<Name extends InputNames> = ReturnType<
 > extends Promise<infer T>
   ? T
   : never;
-
-export async function insertInput<Name extends InputNames>(
-  ctx: MutationCtx,
-  worldId: Id<'worlds'>,
-  name: Name,
-  args: InputArgs<Name>,
-): Promise<Id<'inputs'>> {
-  const worldStatus = await ctx.db
-    .query('worldStatus')
-    .withIndex('worldId', (q) => q.eq('worldId', worldId))
-    .unique();
-  if (!worldStatus) {
-    throw new Error(`World for engine ${worldId} not found`);
-  }
-  return await engineInsertInput(ctx, worldStatus.engineId, name, args);
-}
